Logic moved to RhCheSeleniumClassModule - inspired from upstream.
