+++ functional-tests/src/main/java/com/redhat/che/selenium/core/RhCheSeleniumSuiteModule.java
@@ -11,25 +11,18 @@
  */
 package com.redhat.che.selenium.core;
 
-import static com.google.inject.matcher.Matchers.any;
 import static com.google.inject.name.Names.named;
 
 import com.google.inject.AbstractModule;
-import com.google.inject.Injector;
-import com.google.inject.Provider;
 import com.google.inject.Provides;
-import com.google.inject.TypeLiteral;
 import com.google.inject.assistedinject.FactoryModuleBuilder;
 import com.google.inject.name.Named;
 import com.google.inject.name.Names;
-import com.google.inject.spi.TypeEncounter;
-import com.google.inject.spi.TypeListener;
 import com.google.inject.util.Providers;
 import com.redhat.che.selenium.core.client.RhCheTestWorkspaceServiceClient;
 import com.redhat.che.selenium.core.workspace.ProvidedWorkspace;
 import com.redhat.che.selenium.core.workspace.RhCheTestWorkspaceImpl;
 import com.redhat.che.selenium.core.workspace.RhCheTestWorkspaceProvider;
-import java.lang.reflect.Field;
 import org.eclipse.che.selenium.core.client.TestWorkspaceServiceClient;
 import org.eclipse.che.selenium.core.client.TestWorkspaceServiceClientFactory;
 import org.eclipse.che.selenium.core.configuration.SeleniumTestConfiguration;
@@ -37,9 +30,7 @@ import org.eclipse.che.selenium.core.configuration.TestConfiguration;
 import org.eclipse.che.selenium.core.provider.DefaultTestUserProvider;
 import org.eclipse.che.selenium.core.user.DefaultTestUser;
 import org.eclipse.che.selenium.core.user.MultiUserCheDefaultTestUserProvider;
-import org.eclipse.che.selenium.core.workspace.InjectTestWorkspace;
 import org.eclipse.che.selenium.core.workspace.TestWorkspace;
-import org.eclipse.che.selenium.core.workspace.TestWorkspaceInjector;
 import org.eclipse.che.selenium.core.workspace.TestWorkspaceProvider;
 
 public class RhCheSeleniumSuiteModule extends AbstractModule {
@@ -60,28 +51,6 @@ public class RhCheSeleniumSuiteModule extends AbstractModule {
       bind(String.class)
           .annotatedWith(Names.named("che.workspaceName"))
           .toProvider(Providers.<String>of(null));
-    }
-    bindListener(any(), new RhCheWorkspaceTypeListener(binder().getProvider(Injector.class)));
-  }
-
-  private class RhCheWorkspaceTypeListener implements TypeListener {
-    private final Provider<Injector> injectorProvider;
-
-    public RhCheWorkspaceTypeListener(Provider<Injector> injector) {
-      this.injectorProvider = injector;
-    }
-
-    @Override
-    public <I> void hear(TypeLiteral<I> type, TypeEncounter<I> encounter) {
-      Class<?> clazz = type.getRawType();
-      for (Field field : clazz.getDeclaredFields()) {
-        if (field.getType() == RhCheTestWorkspaceImpl.class
-            && field.isAnnotationPresent(InjectTestWorkspace.class)) {
-          encounter.register(
-              new TestWorkspaceInjector<>(
-                  field, field.getAnnotation(InjectTestWorkspace.class), injectorProvider));
-        }
-      }
     }
   }
